compile warning fix
git-svn-id: http://192.168.20.11/sequoiadb/trunk@36808 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/tools/seadapter/seAdptMgr.cpp
@@ -1551,7 +1551,8 @@ namespace seadapter
                    "Index information query request failed[ %d ]", rc ) ;
 
       // All text indices information should be replied in one object.
-      if ( expectReplySz != numReturned || expectReplySz != objVec.size() )
+      if ( ( expectReplySz != numReturned ) ||
+           ( expectReplySz != (INT32)objVec.size() ) )
       {
          PD_LOG( PDERROR, "Reply of index information is not as expected" ) ;
          rc = SDB_SYS ;
